requirements.txt
alembic==1.13.1
asyncpg==0.29.0
bcrypt==4.0.1
black==23.11.0
fastapi[all]==0.108.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
python-jose==3.3.0
ruff==0.1.9
sqlalchemy==2.0.25